% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{kmer}
\alias{kmer}
\title{Create k-mers}
\usage{
kmer(x, k)
}
\arguments{
\item{x}{A \code{list} of sequences created by}

\item{k}{A length of \code{k-mers}.}
}
\value{
A \code{list} of k-mers.
}
\description{
Creates k-mers from fasta file.
}
\examples{
kmer(read_fasta_f(system.file(package = "ReadFastaKmers", "test.fasta")), 5)
}
