% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.fasta_sequence}
\alias{print.fasta_sequence}
\title{Prints fasta sequence}
\usage{
\method{print}{fasta_sequence}(x, ...)
}
\arguments{
\item{x}{object after read_fasta_f}

\item{...}{other args}
}
\description{
Prints fasta sequence
}
