% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{read_fasta_f}
\alias{read_fasta_f}
\title{Read fasta}
\usage{
read_fasta_f(x)
}
\arguments{
\item{x}{A \code{fasta} file.}
}
\value{
A \code{list} of sequences and their names.
}
\description{
Reads fasta file.
}
\examples{
read_fasta_f(system.file(package = "ReadFastaKmers", "test.fasta"))
}
