% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.fasta_sequence}
\alias{summary.fasta_sequence}
\title{Summary fasta sequence}
\usage{
\method{summary}{fasta_sequence}(object, ...)
}
\arguments{
\item{object}{object after read_fasta_f()}

\item{...}{other args}
}
\description{
Summary fasta sequence
}
