% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.kmer}
\alias{summary.kmer}
\title{Summary kmer}
\usage{
\method{summary}{kmer}(object, ...)
}
\arguments{
\item{object}{object after kmer()}

\item{...}{other args}
}
\description{
Summary kmer
}
